requirements.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.2.3
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-pdf==0.2.0
fpdf==1.7.2
gunicorn==20.1.0
Pillow==8.2.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
pytz==2021.1
sqlparse==0.4.1
whitenoise==5.2.0
